requirements.txt
psycopg2-binary
toml
click
plotly
dash
pandas
openpyxl
scikit-learn
kats